requirements.txt
beautifulsoup4
scikit-learn
pandas
requests
selenium
mlflow[pytorch]
tqdm
Pillow

--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.7.1+cu118
torchvision==0.22.1+cu118
